renamed authController to registerController
