+++ swdapp/src/lib/server/controllers/registerController.ts
